1.Changes RegisterForm (added new fields) 2.Added states,store,rtk etc 3.Connected Login and Register page 4. Adjusted router for multi role app 5.Added Logout button and more..

+++ backend/src/models/User.ts
@@ -32,7 +32,7 @@ User.init({
     role: { type: DataTypes.STRING, defaultValue: 'User' },
     weight: { type: DataTypes.FLOAT, allowNull: false },
     height: { type: DataTypes.FLOAT, allowNull: false },
-    physicalActivity: { type: DataTypes.INTEGER, allowNull: false },
+    physicalActivity: { type: DataTypes.FLOAT, allowNull: false },
 
 }, {
     timestamps: true,
